+++ core/src/codegen/from_type_param.rs
@@ -103,7 +103,7 @@ impl<'a> OuterFromImpl<'a> for FromTypeParamImpl<'a> {
     }
 
     fn trait_bound(&self) -> syn::Path {
-        path!(::darling::FromMetaItem)
+        path!(::darling::FromMeta)
     }
 
     fn base(&'a self) -> &'a TraitImpl<'a> {