+++ core/src/codegen/from_variant_impl.rs
@@ -107,7 +107,7 @@ impl<'a> OuterFromImpl<'a> for FromVariantImpl<'a> {
     }
 
     fn trait_bound(&self) -> syn::Path {
-        path!(::darling::FromMetaItem)
+        path!(::darling::FromMeta)
     }
 
     fn base(&'a self) -> &'a TraitImpl<'a> {