+++ core/src/options/shape.rs
@@ -1,7 +1,7 @@
 use quote::{ToTokens, Tokens};
 use syn::{Meta, NestedMeta};
 
-use {Error, FromMetaItem, Result};
+use {Error, FromMeta, Result};
 
 #[derive(Debug, Clone)]
 pub struct Shape {
@@ -29,7 +29,7 @@ impl Default for Shape {
     }
 }
 
-impl FromMetaItem for Shape {
+impl FromMeta for Shape {
     fn from_list(items: &[NestedMeta]) -> Result<Self> {
         let mut new = Shape::default();
         for item in items {
@@ -142,7 +142,7 @@ impl DataShape {
     }
 }
 
-impl FromMetaItem for DataShape {
+impl FromMeta for DataShape {
     fn from_list(items: &[NestedMeta]) -> Result<Self> {
         let mut new = DataShape::default();
         for item in items {
@@ -207,35 +207,35 @@ mod tests {
     use syn;
 
     use super::Shape;
-    use FromMetaItem;
+    use FromMeta;
 
     /// parse a string as a syn::Meta instance.
-    fn pmi(tokens: Tokens) -> ::std::result::Result<syn::Meta, String> {
+    fn pm(tokens: Tokens) -> ::std::result::Result<syn::Meta, String> {
         let attribute: syn::Attribute = parse_quote!(#[#tokens]);
         attribute.interpret_meta().ok_or("Unable to parse".into())
     }
 
-    fn fmi<T: FromMetaItem>(tokens: Tokens) -> T {
-        FromMetaItem::from_meta_item(&pmi(tokens).expect("Tests should pass well-formed input"))
+    fn fm<T: FromMeta>(tokens: Tokens) -> T {
+        FromMeta::from_meta(&pm(tokens).expect("Tests should pass well-formed input"))
             .expect("Tests should pass valid input")
     }
 
     #[test]
     fn supports_any() {
-        let decl = fmi::<Shape>(quote!(ignore(any)));
+        let decl = fm::<Shape>(quote!(ignore(any)));
         assert_eq!(decl.any, true);
     }
 
     #[test]
     fn supports_struct() {
-        let decl = fmi::<Shape>(quote!(ignore(struct_any, struct_newtype)));
+        let decl = fm::<Shape>(quote!(ignore(struct_any, struct_newtype)));
         assert_eq!(decl.struct_values.any, true);
         assert_eq!(decl.struct_values.newtype, true);
     }
 
     #[test]
     fn supports_mixed() {
-        let decl = fmi::<Shape>(quote!(ignore(struct_newtype, enum_newtype, enum_tuple)));
+        let decl = fm::<Shape>(quote!(ignore(struct_newtype, enum_newtype, enum_tuple)));
         assert_eq!(decl.struct_values.newtype, true);
         assert_eq!(decl.enum_values.newtype, true);
         assert_eq!(decl.enum_values.tuple, true);